Add handlers for split, ejectMass and move commands, send according messages to mechanics, add EmptyReplicator class

+++ src/main/java/model/gameInfo/GameSessionImpl.java
@@ -15,7 +15,7 @@ import java.util.List;
  */
 public class GameSessionImpl implements GameSession {
     private static final SequentialIDGenerator idGenerator = new SequentialIDGenerator();
-    private final int id = idGenerator.next();
+    private final long id = idGenerator.next();
     @NotNull
     private final GameField field = new GameField();
     @NotNull
